refactor(dex): flatten transport module into crate-root re-exports
Address review findings on #742: a pub mod requires a stated reason,
so re-export the swap vocabulary items at the crate root instead and
make the module private. The transfer traits and IBC_TIMEOUT have
impl-side consumers only, so gate them behind the impl feature to
keep non-impl builds free of dead-code errors, and drop the redundant
ungated aliases in favor of the established TransportOut names.
Separate import tiers in transport/mod.rs with a blank line.
